rewrite compile folder and simplify lapikit
