--- src/lib/compiler/plugins.ts
@@ -1,8 +1,0 @@
-const lapikitPlugins = {
-	repl: {
-		components: ['repl'],
-		ref: '@lapikit/repl'
-	}
-} as const;
-
-export default lapikitPlugins;
