--- src/lib/compiler/preprocess/index.ts
@@ -1,1 +1,0 @@
-export * from './lili';
