+++ src/lib/components/accordion/modules/accordion-item.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { useClassName, useStyles } from '$lib/utils';
-	import { makeComponentProps } from '$lib/compiler/mapped-code';
+	import { makeComponentProps } from '$lib/html-mapped';
 	import type { AccordionItemModelProps, AccordionItemProps } from '../accordion.types.ts';
 
 	let {
