+++ src/lib/components/aspect-ratio/aspect-ratio.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { useClassName, useStyles } from '$lib/utils';
-	import { makeComponentProps } from '$lib/compiler/mapped-code';
+	import { makeComponentProps } from '$lib/html-mapped';
 	import type { AspectRatioFit, AspectRatioProps, AspectRatioValue } from './aspect-ratio.types.ts';
 
 	const FALLBACK_RATIO = 16 / 9;
