+++ src/lib/components/separator/separator.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { useClassName, useStyles } from '$lib/utils';
-	import { makeComponentProps } from '$lib/compiler/mapped-code';
+	import { makeComponentProps } from '$lib/html-mapped';
 	import type { SeparatorProps } from './separator.types.ts';
 
 	let {
