+++ src/lib/components/sheet/sheet.svelte
@@ -5,7 +5,7 @@
 	 */
 
 	import { useClassName, useStyles } from '$lib/utils';
-	import { makeComponentProps } from '$lib/compiler/mapped-code';
+	import { makeComponentProps } from '$lib/html-mapped';
 
 	let {
 		class: className,
