+++ src/lib/html-mapped.ts
@@ -1,5 +1,4 @@
-import type { SClassProp, SStyleProp } from '$lib/@types';
-import type { PropValue } from '$lib/@types';
+import type { PropValue, SClassProp, SStyleProp } from '$lib/@types';
 
 /**
  * Computes a string of class names based on the provided sClass and classDirectiveProps.
